[E3DB-959] Test app clean-up.
