+++ androidtest/src/main/res/layout/activity_main.xml
@@ -3,24 +3,27 @@
         xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
         xmlns:tools="http://schemas.android.com/tools"
+        android:id="@+id/container"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         tools:context="com.tozny.e3dbtest.MainActivity">
 
-    <TextView
-            android:id="@+id/hello_label"
-            android:layout_width="wrap_content"
+    <FrameLayout
+            android:id="@+id/content_fragment"
+            android:layout_width="fill_parent"
+            android:layout_height="fill_parent"
+            android:layout_weight="1" />
+
+    <android.support.design.widget.BottomNavigationView
+            android:id="@+id/navigation"
+            android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:text="Hello World!"
+            android:layout_marginEnd="0dp"
+            android:layout_marginStart="0dp"
+            android:background="?android:attr/windowBackground"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toTopOf="parent"/>
-
-    <Button
-            android:id="@+id/register_client_button"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Register Client"/>
+            app:menu="@menu/navigation"/>
 
 </android.support.constraint.ConstraintLayout>
