+++ androidtest/src/main/res/values/strings.xml
@@ -8,7 +8,7 @@
 
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
-    <string name="title_activity_main4">Main4Activity</string>
+    <string name="title_activity_main">MainActivity</string>
     <string name="load_config">Load Config</string>
     <string name="delete_config">Delete Config</string>
     <string name="new_config">New Config</string>
